feat(modding): added values display syntax to quests
Syntax is $rep{id} for reputation, $item{id} for quest items and $comp{id} for components, where id is a database id of required character/item/component. Example is $item{6} for displaying count of item with id 6.
This can be inserted in a middle of other text
This can be used almost anywhere where localized text can be used.